+++ json_temp/Description.json
@@ -1,3 +1,3 @@
 {
-    "Description": "Simon Fraser University (SFU) is a public research university located in British Columbia, Canada. Founded in 1965, it is named after the Scottish explorer Simon Fraser, who played a key role in the province's history. SFU is known for its commitment to research, a diverse range of academic programs, and its focus on community engagement and applied learning.\n\n### Key Features of Simon Fraser University:\n\n- **Campuses**: SFU has three campuses located in British Columbia:\n  1. **Burnaby Campus**: The main campus, situated on Burnaby Mountain, features stunning views and contains most of the university's faculties and academic programs.\n  2. **Vancouver Campus**: Located in the heart of downtown Vancouver, this campus focuses on programs in business, arts, and environmentally sustainable development.\n  3. **Surrey Campus**: This campus offers programs in health science, business, and engineering, emphasizing innovation and community engagement.\n\n- **Academics**: SFU offers a wide range of undergraduate and graduate programs across various disciplines, including arts, sciences, business, engineering, health sciences, education, and more. The university is known for its interdisciplinary approach and is home to several research institutes and centers.\n\n- **Research**: Recognized as one of Canada\u2019s leading research institutions, SFU is heavily involved in various fields, including health, social sciences, engineering, and environmental studies. The university emphasizes innovation and collaboration with industry and community organizations.\n\n- **Community Engagement**: SFU has a strong emphasis on integrating education with community service and outreach. It encourages students to engage with local communities through experiential learning, co-op programs, and volunteering.\n\n- **Campus Life**: The university has a vibrant student life, with numerous clubs, organizations, and recreational activities available. SFU offers sports teams, cultural and artistic events, and various student services to support overall student well-being.\n\n- **Internationalization**: SFU has a diverse student body, attracting students from around the world. The university promotes international learning opportunities and has numerous partnerships with institutions globally.\n\n- **Sustainability**: SFU is recognized for its commitment to sustainability and environmental stewardship, integrating these values into its curriculum, research, and campus practices.\n\nOverall, Simon Fraser University is known for its innovative programs, strong research focus, and commitment to public good, making it a prominent institution within Canadian higher education."
+    "Description": "Google LLC is an American multinational company that specializes in various technology-related services and products. Founded in September 1998 by Larry Page and Sergey Brin while they were Ph.D. students at Stanford University, Google initially started as a search engine. It has since expanded into a vast ecosystem of products and services. Here\u2019s an overview of Google:\n\n### Core Services\n1. **Search Engine**: Google Search is the most widely used search engine globally, known for its efficiency and accurate results based on complex algorithms.\n\n2. **Advertising**: Google generates a significant portion of its revenue through online advertising, primarily through platforms like Google Ads and Google AdSense. These services allow businesses to display ads on search results pages and across partner sites.\n\n3. **Cloud Computing**: Google Cloud offers a suite of cloud services, including data storage, computing power, and machine learning tools, catering to enterprise customers and developers.\n\n4. **Software Products**:\n   - **Android**: An open-source mobile operating system used by a vast majority of smartphones globally.\n   - **Chrome**: A web browser known for its speed and efficiency, often favored for web browsing.\n   - **Gmail**: A widely-used email service offering robust features and huge storage.\n\n5. **Productivity Tools**: Google Workspace (formerly G Suite) includes applications like Google Docs, Sheets, Slides, and Drive, enabling collaboration and productivity for individuals and businesses.\n\n### Hardware\nGoogle has ventured into hardware with products such as:\n- **Pixel Phones**: Smartphones known for their high-quality cameras and integration with Google services.\n- **Google Nest**: A line of smart home products including speakers (Google Nest Audio), displays (Nest Hub), and security cameras.\n\n### Innovations and Research\nGoogle is involved in various research initiatives and innovations, including:\n- **Artificial Intelligence**: Google\u2019s AI research division, Google AI, works on advancements in machine learning, natural language processing, and other AI technologies.\n- **Waymo**: A subsidiary focused on developing self-driving car technology.\n- **Verily**: A life sciences research organization that applies technology to health care.\n\n### Corporate Structure\nIn 2015, Google underwent a corporate restructuring and became a subsidiary of Alphabet Inc., a holding company that encompasses various businesses under the Alphabet umbrella. This restructuring allows Google to focus on its core services while other ventures can operate more independently.\n\n### Social Responsibility and Challenges\nGoogle is also involved in various social responsibility initiatives, including sustainability efforts and digital literacy programs. However, the company faces challenges and controversies related to privacy concerns, antitrust scrutiny, and its impact on news and media.\n\n### Global Presence\nGoogle operates globally, with offices and data centers in many countries. It continues to expand its services and adapt to different markets while maintaining its reputation as a leader in technology innovation.\n\n### Conclusion\nGoogle\u2019s influence on the internet and technology is profound, impacting how people access information, communicate, and interact with digital tools and platforms. Its continuous innovation keeps it at the forefront of the tech industry."
 }
