+++ json_temp/email.json
@@ -1,1 +1,1 @@
-[{"email": "jpei@cs.sfu.ca", "full_name": "Jian Pei"}, {"email": "veronica@cs.sfu.ca", "full_name": "Veronica Dahl"}, {"email": "stella@cs.sfu.ca", "full_name": "Stella Atkins"}, {"email": "funda@cs.sfu.ca", "full_name": "Funda Ergun"}, {"email": "kamal@cs.sfu.ca", "full_name": "Kamal Gupta"}, {"email": "jcliu@cs.sfu.ca", "full_name": "Angie Zhang"}, {"email": "ljilja@cs.sfu.ca", "full_name": "Angie Zhang"}, {"email": "angiez@cs.sfu.ca", "full_name": "Angie Zhang"}, {"email": "tayebi@cs.sfu.ca", "full_name": "Mohammad Tayebi"}, {"email": "mehta@cs.sfu.ca", "full_name": "Manish Mehta"}, {"email": "barbara_hilden@sfu.ca", "full_name": "Barbara Hilden"}, {"email": "gustavo_balbinot@sfu.ca", "full_name": "Gustavo Balbinot"}, {"email": "marilyn_e_trautman@sfu.ca", "full_name": "Marilyn Trautman"}, {"email": "shivani_kundra@sfu.ca", "full_name": "Shivani Kundra"}, {"email": "damon_van_der_linde@sfu.ca", "full_name": "Damon Linde"}, {"email": "sarah_mckay@sfu.ca", "full_name": "Sarah McKay"}, {"email": "kieran_cox@sfu.ca", "full_name": "Kieran Cox"}, {"email": "amr_marzouk@sfu.ca", "full_name": "Amr Marzouk"}, {"email": "alexis_carr@sfu.ca", "full_name": "Alexis Carr"}, {"email": "mkhan@sfu.ca", "full_name": "Madiha Khan"}, {"email": "ian_reddy@ucs.sfu.ca", "full_name": "Ian Reddy"}]
+[]
