+++ json_temp/ip_safe_list.json
@@ -1,26 +1,8 @@
 [
     {
-        "ip": "142.58.142.231"
+        "ip": "172.217.14.238"
     },
     {
-        "ip": "142.58.233.251"
-    },
-    {
-        "ip": "208.70.244.23"
-    },
-    {
-        "ip": "142.58.142.32"
-    },
-    {
-        "ip": "142.58.142.66"
-    },
-    {
-        "ip": "142.58.142.134"
-    },
-    {
-        "ip": "142.58.143.29"
-    },
-    {
-        "ip": "142.58.143.9"
+        "ip": "216.239.32.29"
     }
 ]
